Rename Mixin classes to match vanilla class names
