+++ src/main/java/me/juancarloscp52/entropy/mixin/AbstractSelectionListAccessor.java
@@ -5,7 +5,7 @@ import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.gen.Accessor;
 
 @Mixin(AbstractSelectionList.class)
-public interface EntryListWidgetAccessor {
+public interface AbstractSelectionListAccessor {
 
     @Accessor
     boolean getScrolling();
