+++ src/main/java/me/juancarloscp52/entropy/mixin/MouseHandlerMixin.java
@@ -27,7 +27,7 @@ import org.spongepowered.asm.mixin.injection.ModifyArg;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(MouseHandler.class)
-public class MouseMixin {
+public class MouseHandlerMixin {
 
     @Shadow
     private double accumulatedDX;
