+++ src/main/java/me/juancarloscp52/entropy/mixin/MthMixin.java
@@ -8,7 +8,7 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(Mth.class)
-public class MathHelperMixin {
+public class MthMixin {
     @Inject(method = "lerpInt(FII)I", at = @At("HEAD"), cancellable = true)
     private static void lerp(float delta, int start, int end, CallbackInfoReturnable<Integer> cir) {
         if(Variables.stuttering)
