+++ src/main/java/me/juancarloscp52/entropy/mixin/SoundEngineMixin.java
@@ -17,22 +17,22 @@
 
 package me.juancarloscp52.entropy.mixin;
 
-import me.juancarloscp52.entropy.client.EntropyClient;
-import net.minecraft.client.multiplayer.ClientLevel;
+import me.juancarloscp52.entropy.Variables;
+import net.minecraft.client.resources.sounds.SoundInstance;
+import net.minecraft.client.sounds.SoundEngine;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
-@Mixin(ClientLevel.class)
-public class ClientWorldMixin {
+@Mixin(SoundEngine.class)
+public class SoundEngineMixin {
 
-    @Inject(method = "disconnect", at = @At("HEAD"))
-    private void onDisconnect(CallbackInfo ci) {
-        if (EntropyClient.getInstance().clientEventHandler == null)
-            return;
-        EntropyClient.getInstance().clientEventHandler.endChaos();
-        EntropyClient.getInstance().clientEventHandler = null;
+    @Inject(method = "calculatePitch", at = @At("RETURN"), cancellable = true)
+    private void forcePitch(SoundInstance soundInstance, CallbackInfoReturnable<Float> cir) {
+        if (Variables.forcePitch) {
+            cir.setReturnValue(Variables.forcedPitch);
+        }
     }
 
 }
