+++ src/main/resources/entropy.mixins.json
@@ -4,8 +4,8 @@
   "package": "me.juancarloscp52.entropy.mixin",
   "compatibilityLevel": "JAVA_16",
   "mixins": [
-    "AbstractBlockStateMixin",
     "BlockMixin",
+    "BlockStateBaseMixin",
     "CatInvoker",
     "EntityMixin",
     "FallingBlockEntityAccessor",
@@ -13,25 +13,25 @@
     "WolfInvoker"
   ],
   "client": [
-    "BackgroundRendererMixin",
-    "BossBarHudAccessor",
-    "BossBarHudMixin",
+    "AbstractSelectionListAccessor",
+    "BossHealthOverlayAccessor",
+    "BossHealthOverlayMixin",
     "CameraMixin",
-    "ClientPlayerEntityMixin",
-    "ClientPlayerInteractionManagerAccessor",
-    "ClientWorldMixin",
-    "EntryListWidgetAccessor",
+    "CameraTypeMixin",
+    "ClientLevelMixin",
+    "FogRendererMixin",
     "GameRendererMixin",
-    "InGameOverlayRendererMixin",
     "KeyboardInputMixin",
     "LivingEntityRendererMixin",
-    "MathHelperMixin",
-    "MinecraftClientAccessor",
-    "MouseMixin",
+    "LocalPlayerMixin",
+    "MinecraftAccessor",
+    "MouseHandlerMixin",
+    "MthMixin",
+    "MultiPlayerGameModeAccessor",
     "OptionsScreenMixin",
-    "PerspectiveMixin",
-    "SheepWoolFeatureRendererMixin",
-    "SoundSystemMixin"
+    "ScreenEffectRendererMixin",
+    "SheepFurLayerMixin",
+    "SoundEngineMixin"
   ],
   "injectors": {
     "defaultRequire": 1
